Switch to 64-bit Mersenne Twister.
This fixes regenerating the same random stream from saved seeds.

Also, add a test case for detecting overconstrained state spaces, and
rewrite the list_shrink example so it always shrinks. Previously, it
could return a list of the same length when splitting, leading to an
infinite loop if the bloom filter was disabled.