Significantly expand the utility of the progress_cb.
Instead of just calling the progress callback after a trial, call it in
several contexts, with a `struct theft_progress_info` tagged to indicate
what information is available in its union.

Depending on the context (tagged by `enum theft_progress_callback_type),
the callback's return value can be used to halt the current activity
early, re-run the same step with changes to the testing environment,
etc.