Remove theft_set_seed from public API.
Setting the seed should be managed by the test runner (possibly
based on a seed configured in the theft_run_config struct)
rather than by the user, because setting a new seed also
stops using the current bit pool, if any.

Move its tests into test_theft_prng. There is a regression test
that should be rewritten -- set it to SKIP for now.