Finish implementing autoshrinking.
It seems to work well to align mutations to the PRNG bit pool to the
sizes of the user's bit requests.

Add tests for individual autoshrink steps (with an injected sequence
of "random" bits, to control which tactics are chosen).

Switch from `bool autoshrink` in type info config to
`struct theft_autoshrink_config` with a `bool enable` field.

Make printing of the autoshrink bit pool configurable.

Always drop at least 1 request. (But add a magic value to disable
this, because it complicates tests.)

If a mutation has no effect, try again (up to a retry limit).