Switch to filling autoshrink bit pool on demand.
Profiling indicates that the majority of runtime is spent
eagerly filling the autoshrink random bit pool. Filling it
on demand, in 64-bit increments, leads to significantly
less overall work than pre-filling it. (Its bit buffer
is realloc'd to double the size when necessary.)

Also, stop returning the initial seed used as the first random
value when setting the seed. Update an integration test to
verify that the provided seed is _used_, not returned.