autoshrink: When dropping, zero bits after last request.
This is necessary for dropping to be monotonic -- otherwise, a drop
could cause higher values to be used, and they may not shrink properly
because paths from them to simpler values are already set in the bloom
filter.

This led to cases where autoshrink found a minimal value in the state
space, but then jumped away from it, and got stuck at the other value.