Various logging changes.
Adjust some log levels.

Add log statements to print the the various hashes used for the bloom
filter, and the bits being hashed from the autoshrink bit pool.

Add ("/%d", rem_bits) to the printout of the bit pool and hash,
to indicate when only a subset of a byte is used.