autoshrinking: Copy as much as possible per step when filling buffer.
The bit pool -> bit buffer copy is a hotspot, and previously it
was only copying a single bit at a time. Restructure it so that
the inner loop copies as many bits as possible at each step
(based on the bit offsets for the src and dst buffers). This speeds up
autoshrinking a bit more.