Forking: Add documentation, custom signal behavior.
The timeout handler can now customize the `kill` signal, to trigger
cleanup or other custom test behavior, and a signaled child that
handles the signal and returns `THEFT_TRIAL_PASS` or calls
`exit(EXIT_SUCCESS)` will still pass.

This gives the test the option of inspecting whether its results
should count as a pass, even though it timed out (in case there's
a race on timing out vs. test completion, but it's already past
the section that actually detects failures).