Put run_info on `struct theft`, rather than passing both.
Add `void *theft_hook_get_env(struct theft *t)`.

For `theft_hook_get_env` to work, `struct theft *t` will
need to be able to return `run_info->hooks.env`, and there
are a lot of places where both are passed as arguments.
Save `run_info` on `t` and just pass that around, instead.

These structs could probably be combined in a later refactor; they are
distinct structs because a previous version of the API initialized the
theft handle independently from the run_info so it could be reused, but
that turned out to be pointless.

Either way, `struct theft` is opaque, so that would not
change the public API.