Change to type-safe property function pointers; make arity explicit.
Instead of a single function pointer typedef with unconstrained
arguments, use a set of typedefs with instance argument counts prop1,
prop2, ... prop7.

This was partially motivated by adding the first `struct theft *t`
argument (for use with `theft_hook_get_env`) to property functions,
because that API change would not break the build with a type
mismatch error, and would be especially likely to cause a crash.

Update tests to make the prop arity explicit.