Add a tag field to `print_trial_result_env` & warn about misuse.
If the `trial_post` hook is overridden and calls
`theft_print_trial_result`, check whether the *env argument
for the `print_trial_result_env` is the overall *env for hooks.

If so, print a warning, unless its tag field has been set -- in
this case, it's likely that it's actually a different kind of
custom test environment, but was accidentally passed to
`theft_print_trial_result` or `theft_hook_trial_post_print_result`
(which casts *env and then calls `theft_print_trial_result`)
directly. In that case, the print_env should probably be a field
on the custom environment.