BUGFIX: Don't ignore byte 0 of autoshrink bit pool when truncating.
This caused strange autoshrinking behavior, which was ultimately
caused by truncating the bit pool to zero bytes when there was only
a single bit remaining in the pool, in the second byte. Instead of
shrinking by shifting it into the first byte, it consistently
overshot and turned it to zero.