Add basic logging API.
This makes the logging level part of the internal state of the test
runner, and adds a macro that conditionally logs based on the log
level (to the current log file).

While the hooks make adding this easy, it's a common enough case that it
should be part of the API.

When concurrency is added, this will be changed somewhat (making
log levels and files worker-process-specific, adding buffering, etc.).

Resolves #13.