Make autoshrinking the default, remove `shrink` and `hash` from API.
At this point, autoshrinking is working sufficiently well that it would
be better to make it the default and focus on continuing to improve it,
rather than dividing efforts between it and custom shrinking.

It can be tricky to write custom shrinkers that aren't potentially
quadratic, and autoshrinking already has heuristics to avoid this built
into it.

Supporting custom shrinkers will also interact badly with concurrency
(supporting multiple worker processes for multi-core searching and
shrinking), greatly increasing the cases that will need to be handled
along the way.

If custom shrinkers are removed, then hashing can be handled internally
to autoshrinking, which enables further improvements to hashing that
would otherwise change several details in the public API.

- Eliminate `shrink` and `hash` callbacks from the public API headers
  and documentation.

- Always use autoshrinking and autoshrink's hashing.

- Eliminate a lot of internal branching. (This could use an extra
  cleanup pass before the release.)

- Extract `theft_autoshrink_make_candidate_bit_pool` from
  `theft_autoshrink_shrink`, for testing purposes.

Closes #20.